requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
PyJWT==2.9.0
pydantic-settings==2.6.0
